Removed auto SQL query "select * from beers" and replaced in the graphical user interface with a free text SQL query (write a query in the "search field").
Added controllers for the "homescreen.fxml" and the temporary added "result.xml" so the button on the "home screen" now sends the query via the JDBC connection to the server and stores the data fetched in an object called "develop_SQL_query" and switches stage to the "result" stage.

The "result" stage gets the value from the object "developer_SQL_query" and displays it in the top element text element, this stage also has a "back" button witch makes the user return to the "home screen", when you initialise the "home screen" the "develop_SQL_query" object is resetted.

Didnt look at dynamic resizing so set the "result" stage to 800x600px just like the "home screen".